test: Move throttle pkg under apim pkg
