+++ integration/test-integration/src/test/java/org/wso2/choreo/connect/tests/testCases/apim/throttle/SubscriptionThrottlingTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.choreo.connect.tests.testCases.throttle;
+package org.wso2.choreo.connect.tests.testCases.apim.throttle;
 
 import com.google.common.net.HttpHeaders;
 import org.apache.http.HttpStatus;
@@ -25,22 +25,22 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import org.wso2.am.integration.clients.admin.ApiResponse;
-import org.wso2.am.integration.clients.admin.api.dto.ApplicationThrottlePolicyDTO;
 import org.wso2.am.integration.clients.admin.api.dto.RequestCountLimitDTO;
+import org.wso2.am.integration.clients.admin.api.dto.SubscriptionThrottlePolicyDTO;
 import org.wso2.am.integration.clients.admin.api.dto.ThrottleLimitDTO;
 import org.wso2.am.integration.clients.store.api.v1.dto.ApplicationDTO;
 import org.wso2.am.integration.test.impl.DtoFactory;
 import org.wso2.carbon.automation.test.utils.http.client.HttpResponse;
 import org.wso2.choreo.connect.tests.util.TestConstant;
-import org.wso2.choreo.connect.tests.util.Utils;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 
 import static org.testng.Assert.assertEquals;
 
-public class ApplicationThrottlingTestCase extends ThrottlingBaseTestCase {
-    private ApplicationThrottlePolicyDTO requestCountPolicyDTO;
+public class SubscriptionThrottlingTestCase extends ThrottlingBaseTestCase {
+    private SubscriptionThrottlePolicyDTO requestCountPolicyDTO;
     private final Map<String, String> requestHeaders = new HashMap<>();
     String endpointURL;
     long requestCount = 15L;
@@ -54,24 +54,27 @@ public class ApplicationThrottlingTestCase extends ThrottlingBaseTestCase {
         String policyTimeUnit = "min";
         Integer policyUnitTime = 1;
         String policyDispName = "15PerMin";
-        String policyDesc = "This is a test application throttle policy";
+        String policyDesc = "This is a test subscription throttle policy";
         RequestCountLimitDTO reqCountLimit =
                 DtoFactory.createRequestCountLimitDTO(policyTimeUnit, policyUnitTime, requestCount);
         ThrottleLimitDTO defaultLimit =
                 DtoFactory.createThrottleLimitDTO(ThrottleLimitDTO.TypeEnum.REQUESTCOUNTLIMIT, reqCountLimit, null);
         requestCountPolicyDTO = DtoFactory
-                .createApplicationThrottlePolicyDTO(policyName, policyDispName, policyDesc, false, defaultLimit);
+                .createSubscriptionThrottlePolicyDTO(policyName, policyDispName, policyDesc, false, defaultLimit,-1,
+                        -1, 100, "min", new ArrayList<>(),
+                        true, "", 0);
 
-        // Add the application throttling policy
-        ApiResponse<ApplicationThrottlePolicyDTO> addedPolicy =
-                restAPIAdmin.addApplicationThrottlingPolicy(requestCountPolicyDTO);
+        // Add the subscription throttling policy
+        ApiResponse<SubscriptionThrottlePolicyDTO> addedPolicy =
+                restAPIAdmin.addSubscriptionThrottlingPolicy(requestCountPolicyDTO);
+        Assert.assertEquals(addedPolicy.getStatusCode(), HttpStatus.SC_CREATED);
         requestCountPolicyDTO = addedPolicy.getData();
 
         // creating the application
         ApplicationDTO app = new ApplicationDTO();
-        app.setName("AppThrottlingApp");
-        app.setDescription("Test Application for Application Throttling");
-        app.setThrottlingPolicy(policyName);
+        app.setName("SubThrottlingApp");
+        app.setDescription("Test Application for Subscription Throttling");
+        app.setThrottlingPolicy(TestConstant.APPLICATION_TIER.UNLIMITED);
         app.setTokenType(ApplicationDTO.TokenTypeEnum.JWT);
         ApplicationCreationResponse appResponse = createApplicationWithKeys(app, restAPIStore);
         Assert.assertNotNull(appResponse.getApplicationId(), "Application ID can't be null");
@@ -83,28 +86,26 @@ public class ApplicationThrottlingTestCase extends ThrottlingBaseTestCase {
         requestHeaders.put(TestConstant.AUTHORIZATION_HEADER, "Bearer " + accessToken);
         requestHeaders.put(HttpHeaders.CONTENT_TYPE, "application/json");
 
-        String apiId = createThrottleApi(TestConstant.API_TIER.UNLIMITED, TestConstant.API_TIER.UNLIMITED,
+        String apiId = createThrottleApi(policyName, TestConstant.API_TIER.UNLIMITED,
                 TestConstant.API_TIER.UNLIMITED);
         // get a predefined api request
         endpointURL = getThrottleAPIEndpoint();
-        endpointURL = Utils.getServiceURLHttps(TestConstant.SAMPLE_API_CONTEXT + "/1.0.0/pet/findByStatus");
 
-        HttpResponse subscriptionResponse = subscribeToAPI(apiId, applicationId,
-                TestConstant.SUBSCRIPTION_TIER.UNLIMITED, restAPIStore);
+        HttpResponse subscriptionResponse = subscribeToAPI(apiId, applicationId, policyName, restAPIStore);
         assertEquals(subscriptionResponse.getResponseCode(), HttpStatus.SC_OK, "Failed to subscribe to the API");
         // this is to wait until policy deployment is complete in case it didn't complete already
         Thread.sleep(TestConstant.DEPLOYMENT_WAIT_TIME);
     }
 
-    @Test(description = "Test application throttling")
-    public void testApplicationLevelThrottling() throws Exception {
+    @Test(description = "Test Subscription throttling")
+    public void testSubscriptionLevelThrottling() throws Exception {
         Assert.assertTrue(isThrottled(endpointURL, requestHeaders, null, requestCount),
-                "Request not throttled by request count condition in application tier");
+                "Request not throttled by request count condition in subscription tier");
     }
 
     @AfterClass(alwaysRun = true)
     public void destroy() throws Exception {
         super.cleanUp();
-        restAPIAdmin.deleteApplicationThrottlingPolicy(requestCountPolicyDTO.getPolicyId());
+        restAPIAdmin.deleteSubscriptionThrottlingPolicy(requestCountPolicyDTO.getPolicyId());
     }
 }
