+++ integration/test-integration/src/test/resources/testng.xml
@@ -70,11 +70,11 @@
     <test name="mgw-micro-with-apim" preserve-order="true" parallel="false">
         <classes>
             <class name="org.wso2.choreo.connect.tests.testCaseBefore.APIManagerInitializeTestCase"/>
-            <class name="org.wso2.choreo.connect.tests.testCases.throttle.AdvanceThrottlingTestCase"/>
-            <class name="org.wso2.choreo.connect.tests.testCases.throttle.ApplicationThrottlingTestCase"/>
-            <class name="org.wso2.choreo.connect.tests.testCases.throttle.SubscriptionThrottlingTestCase"/>
+            <class name="org.wso2.choreo.connect.tests.testCases.apim.throttle.AdvanceThrottlingTestCase"/>
+            <class name="org.wso2.choreo.connect.tests.testCases.apim.throttle.ApplicationThrottlingTestCase"/>
+            <class name="org.wso2.choreo.connect.tests.testCases.apim.throttle.SubscriptionThrottlingTestCase"/>
 <!--            TODO: (Praminda) this requires upgrade in apim throttle dependency. enable after upgrade-->
-<!--            <class name="org.wso2.choreo.connect.tests.testCases.throttle.DenyConditionTestCase"/>-->
+<!--            <class name="org.wso2.choreo.connect.tests.testCases.apim.throttle.DenyConditionTestCase"/>-->
             <class name="org.wso2.choreo.connect.tests.testCases.apim.SubscriptionValidationTestCase"/>
             <class name="org.wso2.choreo.connect.tests.testCases.apim.BlockedAPITestCase"/>
         </classes>
